+++ angular-frontend/src/app/omnai-datasource/omnai-scope-server/devicelist.component.ts
@@ -2,15 +2,20 @@
 
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { OmnAIScopeDataService } from './live-data.service';
+import { MatCardModule } from '@angular/material/card';
+import { DeviceCardComponent } from '../../sidebar/devices/devicecard.component';
+import { SideBarService } from '../../sidebar/sidebar.service';
 
 @Component({
     selector: 'app-device-list',
     templateUrl: './devicelist.component.html',
-    imports: [],
+    styleUrl: './devicelist.component.css',
+    imports: [MatCardModule, DeviceCardComponent],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DeviceListComponent {
     private readonly deviceHandler = inject(OmnAIScopeDataService);
     protected readonly devices = this.deviceHandler.devices;
+    readonly sideBarService = inject(SideBarService)
     getDevicesList = this.deviceHandler.getDevices.bind(this.deviceHandler);
 }