+++ angular-frontend/src/app/toolbar/toolbar.component.ts
@@ -4,12 +4,13 @@ import { StartDataButtonComponent } from "../toolbar/start-data-button.component
 import { SaveDataButtonComponent } from '../toolbar/save-data-button.component';
 import { SettingsMenuComponent } from '../settings/setting-menu.component';
 import { MatSlideToggle } from '@angular/material/slide-toggle';
+import { SideBarComponent } from "../sidebar/sidebar.component";
 
 @Component({
     selector: 'app-toolbar',
     templateUrl: './toolbar.component.html',
     styleUrl: './toolbar.component.css',
-    imports: [StartDataButtonComponent, SaveDataButtonComponent, SettingsMenuComponent, MatSlideToggle]
+    imports: [StartDataButtonComponent, SaveDataButtonComponent, SettingsMenuComponent, MatSlideToggle, SideBarComponent]
 })
 export class ToolbarComponent {
     protected readonly advancedMode = inject(AdvancedModeService);